// frontend/src/app/models/rezervacija.ts
export class Rezervacija{
    idRez:number=0;
    gost: string="";
    restoran: string="";
    adresa: string="";
    datumVreme: string=""
    datumVremeOdlaska: string=""
    brojLjudi: number=0;
    dodatniZahtevi:string="";
    status:string=""
    idStola:number=0;
    komentar:string=""
    ocena:number=0
    konobar:string=""
}